Make dep8 nm test directly invokable
By renaming the test to nm.py, it can be called directly with:

    python3 -m unittest nm

To make this work, we also need to move required test setup and teardown
code to setUpModule() and tearDownModule() accordingly, and move other
"script" code into the __name__ == '__main__' idiom.

This then allows easier debugging of failing tests by calling unittest
to run a specific test using the usual method. For example:

    sudo python3 -m unittest nm.ColdplugWifi.test_open_b_ip6_raonly_no_pe