d/t/nm.py: adapt to changes dnsmasq behavior (LP: #1894619)
As described and discussed in (LP: #1894619) the behavior of dnsmasq
has changes. After reverting that for a while upstream confirmed
that - at least for the configuration our test uses - this is the
intended and wanted behavior.

Due to that dnsmasq now (Impish) follows upstream and therefore
the test in NM needs to be adapted to tolerate that new behavior.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>